Restore fmaxf function

+++ components/schemes/Color/src/MDCLegacyColorScheme.m
@@ -93,7 +93,7 @@
   if ([color getHue:&hue saturation:&saturation brightness:&brightness alpha:&alpha]) {
     return [UIColor colorWithHue:hue
                       saturation:saturation
-                      brightness:(CGFloat)fmax(brightness - 0.2, 0)
+                      brightness:(CGFloat)fmaxf((float)brightness - (float)0.2, 0)
                            alpha:alpha];
   }
   return nil;
